requirements.txt
bs4
nltk
gensim
sklearn
wikipedia-1.4.0